Fix ActiveStorage integration.
Render attachments of hash including signed_id and url. Consider a
configuration to hide the signed_id.
Implement a workaround for Rails' broken strong params, which does not
allow permitting an array containing scalars and hashes (needed for when
API consumer must pass scalar signed_ids along with hashes that conform
to the ActiveStorage API.